scripts/rendering/pyproject.toml
[tool.poetry]
name = "rendering"
version = "0.1.0"
description = ""
authors = ["pollenjp <polleninjp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.24.1"
omegaconf = "^2.3.0"
pillow = "^9.4.0"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"
autoflake8 = "^0.4.0"
isort = "^5.11.4"
mypy = "^0.991"
nox = "^2022.11.21"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
